# Faster point sorting + performance tuning