--- a/src/NonuniformFFTs.jl
+++ b/src/NonuniformFFTs.jl
@@ -53,10 +53,6 @@ default_block_size(::Dims, ::GPU) = 1024  # except in 2D and 3D (see below)
 default_block_size(::Dims{2}, ::GPU) = (32, 32)
 default_block_size(::Dims{3}, ::GPU) = (16, 16, 4)  # tuned on A100 with 256³ non-oversampled grid, σ = 2 and m = HalfSupport(4)
 
-# This is used at several places instead of getindex (inside of a `map`) to make sure that
-# the @inbounds is applied.
-inbounds_getindex(v, i) = @inbounds v[i]
-
 function default_workgroupsize(backend, ndrange::Dims)
     # This currently assumes 1024 available threads, which might fail in some GPUs (AMD?).
     KA.default_cpu_workgroupsize(ndrange)
@@ -64,6 +60,10 @@ end
 
 # Case of 1D kernels on the GPU (typically, kernels which iterate over non-uniform points).
 default_workgroupsize(::GPU, ndrange::Dims{1}) = (1024,)
+
+# This is used at several places instead of getindex (inside of a `map`) to make sure that
+# the @inbounds is applied.
+inbounds_getindex(v, i) = @inbounds v[i]
 
 include("blocking.jl")
 include("plan.jl")