We now avoid using `sortperm!` to perform spatial sorting of non-uniform points. This greatly increases performance of `set_points!` on CPU and GPU.

On the GPU, type-2 transforms of complex data now have comparable performance to CuFINUFFT, while type-1 transforms are only about 20% slower (on an A100, see code below for details). Spreading improvements are needed to close this gap in performance.

Note that for real-to-complex transforms we're considerably faster than CuFINUFFT since the latter doesn't support real non-uniform data.

Raw test code with timings:

```julia
using NonuniformFFTs
using Adapt
using CUDA
using FFTW
using LinearAlgebra: norm
using KernelAbstractions
using ThreadPinning
using Random: randn!, Xoshiro
using StaticArrays: SVector
using BenchmarkTools

bench::Bool = true

pinthreads(:affinitymask)

# Note: this currently doesn't work with CUDA (due to the RNG), so it must be done with CPU arrays.
function random_points!(rng, xp::AbstractVector{<:SVector})
    # All types (V, T, N) need to be *inside* of the do-block for things to work with CUDA.
    map!(xp, xp) do x
        V = typeof(x)
        T = eltype(V)
        T(2π) * rand(rng, V)
        # T(π) * (rand(rng, V) .+ T(0.5))  # in [π/2, 3π/2)  // try to avoid periodic folding issues (artificially)
    end
end

##

backend = CUDABackend()
# backend = CPU()

Ns = (256, 256, 256)  # number of Fourier modes in each direction
Np = prod(Ns)    # number of non-uniform points

# Generate some non-uniform random data
T = Complex{Float64}  # could also be e.g. Float64 for real non-uniform data
Tr = real(T)
D = length(Ns)                   # number of dimensions
rng = Xoshiro(42)
xp_init = allocate(CPU(), SVector{D, Tr}, Np)
random_points!(rng, xp_init)
# xp_init = sort_points_by_hilbert_index(xp_init, 64; subsort = false)
xp = adapt(backend, xp_init)  # copy to GPU if needed

xp_tup = ntuple(i -> getindex.(xp, i), D)  # layout needed by CuFINUFFT

vp_init = similar(xp_init, T)
randn!(rng, vp_init)  # random values at points
vp = adapt(backend, vp_init)

@. vp = real(vp)  # for comparing with real-data transforms

##

@show CUDA.attribute(device(), CUDA.DEVICE_ATTRIBUTE_MAX_SHARED_MEMORY_PER_BLOCK) >> 10
@show CUDA.attribute(device(), CUDA.DEVICE_ATTRIBUTE_MAX_SHARED_MEMORY_PER_BLOCK_OPTIN) >> 10
@show CUDA.attribute(device(), CUDA.DEVICE_ATTRIBUTE_MAX_SHARED_MEMORY_PER_MULTIPROCESSOR) >> 10

# Create plan for data of type T.
# Parameters correspond to 1e-7 accuracy.
# Note: sort_points ("sorting") is a bit misleading since it means permuting non-uniform
# locations.
params = (;
    block_size = (16, 16, 4),      # set to `nothing` for default behaviour before PR30 (no spatial sorting; slower)
    sort_points = False(),  # doesn't change much (and needs a couple of extra GPU allocations)
)
p = PlanNUFFT(T, Ns; m = HalfSupport(4), σ = 2.0, backend, fftw_flags = FFTW.ESTIMATE, params...);

CUDA.@time set_points!(p, xp_tup);
# CUDA.@profile trace=true set_points!(p, xp_tup)
if bench
    print("set_points!: "); @btime CUDA.@sync set_points!($p, $xp_tup);   # 4.8ms / 3.9ms (points far from boundaries)
end;

us = KernelAbstractions.zeros(backend, complex(T), size(p));

CUDA.@time exec_type1!(us, p, vp);
if bench
    print("exec_type1!: "); @btime CUDA.@sync exec_type1!($us, $p, $vp);  # 324ms / 315ms (points far from boundaries)
end;

# Impose Hermitian symmetry (for comparison with r2c transforms)
iseven(size(us, 1)) && @views us[end ÷ 2 + 1, :, :] .= 0;
iseven(size(us, 2)) && @views us[:, end ÷ 2 + 1, :] .= 0;
iseven(size(us, 3)) && @views us[:, :, end ÷ 2 + 1] .= 0;

wp = similar(vp);

CUDA.@time exec_type2!(wp, p, us);  # 0.11s (blocking + sorting)
if bench
    print("exec_type2!: "); @btime CUDA.@sync exec_type2!($wp, $p, $us);  # 106ms / 85ms (points far from boundaries)
end;

# Copy results to the CPU for making comparisons later.
us_ours = Array(us);
wp_ours = Array(wp);

## Test r2c transforms

vp_real = real.(vp)
p_real = PlanNUFFT(eltype(vp_real), Ns; m = HalfSupport(4), σ = 2.0, backend, fftw_flags = FFTW.ESTIMATE, params...);

CUDA.@time set_points!(p_real, xp_tup);
# CUDA.@profile trace=true set_points!(p_real, xp_tup)
if bench
    print("set_points! (r2c): ");
    @btime CUDA.@sync set_points!($p_real, $xp_tup);   # 5.4ms
end;

us_r2c = KernelAbstractions.zeros(backend, complex(T), size(p_real));

CUDA.@time exec_type1!(us_r2c, p_real, vp_real);
if bench
    print("exec_type1! (r2c): ")
    @btime CUDA.@sync exec_type1!($us_r2c, $p_real, $vp_real);  # 175ms
end;

# Impose Hermitian symmetry (for comparison with c2c transforms)
@views us_r2c[end, :, :] .= 0;
iseven(size(us_r2c, 2)) && @views us_r2c[:, end ÷ 2 + 1, :] .= 0;
iseven(size(us_r2c, 3)) && @views us_r2c[:, :, end ÷ 2 + 1] .= 0;

wp_real = similar(vp_real);

CUDA.@time exec_type2!(wp_real, p_real, us_r2c);  # !! c2r FFT allocates, https://github.com/JuliaGPU/CUDA.jl/issues/2249
if bench
    print("exec_type2! (r2c): ")
    @btime CUDA.@sync exec_type2!($wp_real, $p_real, $us_r2c);  # 83ms
end;

@show norm(wp - wp_real) / norm(wp)  # 4.9e-16 (good!)

## Compare c2c results with CuFINUFFT

using FINUFFT

opts = (
    modeord = 1,
    upsampfac = 2.0,
    gpu_device_id = CUDA.deviceid(CUDA.device()),
    gpu_stream = Base.unsafe_convert(Ptr{CUDA.CUstream_st}, CUDA.stream()),
    gpu_method = 1,  # 1: GM(-sort) method -- method 0 or 2 is much slower, at least with Float64 data
    gpu_sort = 1,    # needed for good performance
    gpu_kerevalmeth = 1,  # setting it to 0 (direct evaluation) doesn't make a big difference
)

ntrans = 1
n_modes = collect(Ns)
tol = 1e-7
plan_type1 = cufinufft_makeplan(1, n_modes, -1, ntrans, tol; dtype = Tr, opts...)
plan_type2 = cufinufft_makeplan(2, n_modes, +1, ntrans, tol; dtype = Tr, opts...)
finalizer(cufinufft_destroy!, plan_type1)
finalizer(cufinufft_destroy!, plan_type2)

CUDA.@time cufinufft_setpts!(plan_type1, xp_tup...);
CUDA.@time cufinufft_setpts!(plan_type2, xp_tup...);

if us isa CuArray
    if bench
        print("CuFINUFFT setpts: ")
        @btime CUDA.@sync cufinufft_setpts!($plan_type1, $xp_tup...);  # 2.3ms with sorting
    end
    CUDA.@time cufinufft_exec!(plan_type1, vp, us);
    if bench
        print("CuFINUFFT type1: ")
        @btime CUDA.@sync cufinufft_exec!($plan_type1, $vp, $us);  # 278ms with sorting
    end
    # Impose Hermitian symmetry (for comparison with r2c transforms)
    iseven(size(us, 1)) && @views us[end ÷ 2 + 1, :, :] .= 0;
    iseven(size(us, 2)) && @views us[:, end ÷ 2 + 1, :] .= 0;
    iseven(size(us, 3)) && @views us[:, :, end ÷ 2 + 1] .= 0;
    CUDA.@time cufinufft_exec!(plan_type2, us, wp);
    if bench
        print("CuFINUFFT type2: ")
        @btime CUDA.@sync cufinufft_exec!($plan_type2, $us, $wp);  # 106ms with sorting
    end
end

us_finu = Array(us);
wp_finu = Array(wp);

##

@show norm(us_ours - us_finu) / norm(us_ours)  # 1.53e-7
@show norm(wp_ours - wp_finu) / norm(wp_ours)  # 1.57e-7
```